// src/app/api/generated/model/indicador-response.ts
/**
 * AKINE API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * Un indicador del resumen
 */
export interface IndicadorResponse { 
    /**
     * Cantidad, cuando el indicador cuenta cosas
     */
    cantidad?: number;
    /**
     * Identificador estable
     */
    clave?: string;
    /**
     * Texto legible
     */
    etiqueta?: string;
    /**
     * Importe, cuando el indicador suma plata. Decimal exacto
     */
    importe?: number;
    /**
     * Moneda ISO 4217, presente solo cuando hay importe
     */
    moneda?: string;
}