// src/app/api/generated/model/seccion-omitida-response.ts
/**
 * AKINE API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * Seccion no devuelta por falta de permiso
 */
export interface SeccionOmitidaResponse { 
    /**
     * Codigo de permiso que falta
     */
    permisoRequerido?: string;
    /**
     * Nombre de la seccion omitida
     */
    seccion?: string;
}